Fixes bug #396
