+++ build.properties
@@ -1,1 +1,1 @@
-scriptcraft-version=3.3.0
+scriptcraft-version=3.4.0
