+++ src/main/js/lib/task-bukkit.js
@@ -2,11 +2,13 @@
 /*global __plugin, module, server*/
 function bukkitSetTimeout(callback, delayInMillis) {
   var delay = Math.ceil(delayInMillis / 50);
-  var Run = Java.type("java.lang.Runnable");
+  var Run = Java.type('java.lang.Runnable');
   var MyRun = Java.extend(Run, {
     run: callback
   });
-  var task = server.scheduler.runTaskLater(__plugin, new MyRun(), delay);
+  var task = server.scheduler[
+    'runTaskLater(org.bukkit.plugin.Plugin, java.lang.Runnable ,long)'
+  ](__plugin, callback, delay);
   return task;
 }
 function bukkitClearTimeout(task) {
@@ -14,7 +16,9 @@ function bukkitClearTimeout(task) {
 }
 function bukkitSetInterval(callback, intervalInMillis) {
   var delay = Math.ceil(intervalInMillis / 50);
-  var task = server.scheduler.runTaskTimer(__plugin, callback, delay, delay);
+  var task = server.scheduler[
+    'runTaskTimer(org.bukkit.plugin.Plugin, java.lang.Runnable ,long, long)'
+  ](__plugin, callback, delay, delay);
   return task;
 }
 function bukkitClearInterval(bukkitTask) {
